Migrate to onStateChanged function
