+++ example/lib/indicators/check_mark_indicator.dart
@@ -17,7 +17,6 @@ class CheckMarkIndicator extends StatefulWidget {
 class _CheckMarkIndicatorState extends State<CheckMarkIndicator>
     with SingleTickerProviderStateMixin {
   static const _indicatorSize = 150.0;
-  final _helper = IndicatorStateHelper();
 
   /// Whether to render check mark instead of spinner
   bool _renderCompleteState = false;
@@ -31,6 +30,20 @@ class _CheckMarkIndicatorState extends State<CheckMarkIndicator>
       onRefresh: () => Future.delayed(const Duration(seconds: 2)),
       child: widget.child,
       completeStateDuration: const Duration(seconds: 2),
+      onStateChanged: (change) {
+        /// set [_renderCompleteState] to true when controller.state become completed
+        if (change.didChange(to: IndicatorState.complete)) {
+          setState(() {
+            _renderCompleteState = true;
+          });
+
+          /// set [_renderCompleteState] to false when controller.state become idle
+        } else if (change.didChange(to: IndicatorState.idle)) {
+          setState(() {
+            _renderCompleteState = false;
+          });
+        }
+      },
       builder: (
         BuildContext context,
         Widget child,
@@ -41,8 +54,6 @@ class _CheckMarkIndicatorState extends State<CheckMarkIndicator>
             AnimatedBuilder(
               animation: controller,
               builder: (BuildContext context, Widget? _) {
-                _helper.update(controller.state);
-
                 if (controller.scrollingDirection == ScrollDirection.reverse &&
                     prevScrollDirection == ScrollDirection.forward) {
                   controller.stopDrag();
@@ -50,14 +61,6 @@ class _CheckMarkIndicatorState extends State<CheckMarkIndicator>
 
                 prevScrollDirection = controller.scrollingDirection;
 
-                /// set [_renderCompleteState] to true when controller.state become completed
-                if (_helper.didStateChange(to: IndicatorState.complete)) {
-                  _renderCompleteState = true;
-
-                  /// set [_renderCompleteState] to false when controller.state become idle
-                } else if (_helper.didStateChange(to: IndicatorState.idle)) {
-                  _renderCompleteState = false;
-                }
                 final containerHeight = controller.value * _indicatorSize;
 
                 return Container(
