+++ lib/custom_refresh_indicator.dart
@@ -1,5 +1,6 @@
 library custom_refresh_indicator;
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
@@ -8,3 +9,4 @@ part 'src/helpers/positioned_indicator_container.dart';
 part 'src/custom_refresh_indicator.dart';
 part 'src/controller.dart';
 part 'src/helpers/indicator_state_helper.dart';
+part 'src/helpers/indicator_state_change.dart';
